xss_analyzer/requirements/production.txt
fastapi [standard]==0.112.2
clickhouse-connect==0.7.19
pydantic-settings==2.4.0
beautifulsoup4==4.12.3